Make OSX fix darwin specific
It turns out that the fix to build or1k-src on OSX broke building on
other setups. The reason seems to be that on those systems, intl isn't built
and the link thus fails with a "ld: cannot find -lintl".
The real fix would probably be to check if intl is being built, and if
so apply the previous fix.
But until this is properly sorted out, only apply the fix for darwin
host builds.